+++ conventions-vclib/src/main/kotlin/VcLibConventions.kt
@@ -50,6 +50,7 @@ inline fun KotlinDependencyHandler.commonImplementationDependencies() {
     implementation(project.ktor("http"))
     implementation(project.napier())
     implementation(project.ktor("utils"))
+    implementation("net.orandja.obor:obor:${VcLibVersions.obor}")
 }
 
 /**